Add support for asynchronous component updates
This change enables a status bar component to be updated by sending an
asynchronous signal to the process.  A typical use-case is the
updating of a volume compenent in response to the pressing of the
volume up/down key.

This feature uses real-time signals. Thus, the signal number specified
in config.h must be the offset into the real-time signal numbers —
i.e., one specifies n, where SIGRTMIN+n is the actual signal number.